server/requirements.txt
fastapi
uvicorn
pydantic
requests
psycopg2-binary
sqlalchemy
asyncpg
python-dotenv
